Chatbot/requirements.txt
flask
pinecone
langchain
langchain-google-genai
google-generativeai
werkzeug
python-dotenv
pypdf
tiktoken